move inner private class to another file
